Simplified three.asy; added oblique projection and ^^ operator.
Added Metapost --- operator; allow :: to accept a direction specifier.

Allow user-defined ++ binary operator (for consistency with -- binary
operator).

Minor knot velocity optimizations.

Added surface example to documentation; updated cube example.

Updated documentation; fixed cxx warning messages.

+++ frame.h
@@ -38,7 +38,7 @@ public:
   }
 
   int size() {
-    return 1+numFormals+numLocals;
+    return (int) (1+numFormals+numLocals);
   }
 
   access *accessFormal(size_t index) {
